ISPN-10557 extended-statistics org.infinispan.stats package overlaps with core
* move tests also (stats -> extendedstats)
* optimized imports after package rename
* fix package-info
* fix some minor issues: generics, unthrown exceptions, deprecated syntax
* made some inner classes static, where it was possible
