+++ extended-statistics/src/main/java/org/infinispan/extendedstats/Module.java
@@ -5,6 +5,6 @@ import org.infinispan.factories.annotations.InfinispanModule;
 /**
  * {@code InfinispanModule} annotation is required for component annotation processing
  */
-@InfinispanModule(name = "extended-statistics")
+@InfinispanModule(name = "extended-statistics", requiredModules = "core")
 public class Module implements org.infinispan.lifecycle.ModuleLifecycle {
 }
