+++ extended-statistics/src/main/java/org/infinispan/extendedstats/package-info.java
@@ -1,4 +1,4 @@
 /**
- * CLI interpreter engine
+ * Infinispan Extended Statistics.
  */
 package org.infinispan.extendedstats;
