+++ extended-statistics/src/test/java/org/infinispan/extendedstats/BaseClusterTopKeyTest.java
@@ -1,4 +1,4 @@
-package org.infinispan.stats;
+package org.infinispan.extendedstats;
 
 import static org.infinispan.distribution.DistributionTestHelper.addressOf;
 import static org.infinispan.test.TestingUtil.k;
@@ -17,8 +17,8 @@ import org.infinispan.Cache;
 import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.distribution.DistributionManager;
-import org.infinispan.interceptors.impl.TxInterceptor;
 import org.infinispan.extendedstats.topK.CacheUsageInterceptor;
+import org.infinispan.interceptors.impl.TxInterceptor;
 import org.infinispan.test.TestingUtil;
 import org.infinispan.test.fwk.CleanupAfterTest;
 import org.infinispan.util.concurrent.IsolationLevel;
