+++ extended-statistics/src/test/java/org/infinispan/extendedstats/logic/SyncDistTxClusterExtendedStatisticLogicTest.java
@@ -1,14 +1,14 @@
-package org.infinispan.stats.logic;
+package org.infinispan.extendedstats.logic;
 
 import org.infinispan.configuration.cache.CacheMode;
-import org.infinispan.stats.BaseTxClusterExtendedStatisticLogicTest;
+import org.infinispan.extendedstats.BaseTxClusterExtendedStatisticLogicTest;
 import org.testng.annotations.Test;
 
 /**
  * @author Pedro Ruivo
  * @since 6.0
  */
-@Test(groups = "functional", testName = "stats.logic.SyncDistTxClusterExtendedStatisticLogicTest")
+@Test(groups = "functional", testName = "extendedstats.logic.SyncDistTxClusterExtendedStatisticLogicTest")
 public class SyncDistTxClusterExtendedStatisticLogicTest extends BaseTxClusterExtendedStatisticLogicTest {
 
    public SyncDistTxClusterExtendedStatisticLogicTest() {
