+++ extended-statistics/src/test/java/org/infinispan/extendedstats/simple/SyncDistExtendedStatisticTest.java
@@ -1,14 +1,14 @@
-package org.infinispan.stats.simple;
+package org.infinispan.extendedstats.simple;
 
 import org.infinispan.configuration.cache.CacheMode;
-import org.infinispan.stats.BaseNonTotalOrderClusteredExtendedStatisticsTest;
+import org.infinispan.extendedstats.BaseNonTotalOrderClusteredExtendedStatisticsTest;
 import org.testng.annotations.Test;
 
 /**
  * @author Pedro Ruivo
  * @since 6.0
  */
-@Test(groups = "functional", testName = "stats.simple.SyncDistExtendedStatisticTest")
+@Test(groups = "functional", testName = "extendedstats.simple.SyncDistExtendedStatisticTest")
 public class SyncDistExtendedStatisticTest extends BaseNonTotalOrderClusteredExtendedStatisticsTest {
 
    public SyncDistExtendedStatisticTest() {
