+++ extended-statistics/src/test/java/org/infinispan/extendedstats/simple/TotalOrderSyncDistExtendedStatisticTest.java
@@ -1,14 +1,14 @@
-package org.infinispan.stats.simple;
+package org.infinispan.extendedstats.simple;
 
 import org.infinispan.configuration.cache.CacheMode;
-import org.infinispan.stats.BaseTotalOrderClusteredExtendedStatisticsTest;
+import org.infinispan.extendedstats.BaseTotalOrderClusteredExtendedStatisticsTest;
 import org.testng.annotations.Test;
 
 /**
  * @author Pedro Ruivo
  * @since 6.0
  */
-@Test(groups = "functional", testName = "stats.simple.TotalOrderSyncDistExtendedStatisticTest")
+@Test(groups = "functional", testName = "extendedstats.simple.TotalOrderSyncDistExtendedStatisticTest")
 public class TotalOrderSyncDistExtendedStatisticTest extends BaseTotalOrderClusteredExtendedStatisticsTest {
 
    public TotalOrderSyncDistExtendedStatisticTest() {
