+++ extended-statistics/src/test/java/org/infinispan/extendedstats/topK/DistTopKeyTest.java
@@ -1,14 +1,14 @@
-package org.infinispan.stats.topK;
+package org.infinispan.extendedstats.topK;
 
 import org.infinispan.configuration.cache.CacheMode;
-import org.infinispan.stats.BaseClusterTopKeyTest;
+import org.infinispan.extendedstats.BaseClusterTopKeyTest;
 import org.testng.annotations.Test;
 
 /**
  * @author Pedro Ruivo
  * @since 6.0
  */
-@Test(groups = "functional", testName = "stats.topK.DistTopKeyTest")
+@Test(groups = "functional", testName = "extendedstats.topK.DistTopKeyTest")
 public class DistTopKeyTest extends BaseClusterTopKeyTest {
    public DistTopKeyTest() {
       super(CacheMode.DIST_SYNC, 2);
