+++ extended-statistics/src/test/java/org/infinispan/extendedstats/topK/ReplTopKeyTest.java
@@ -1,14 +1,14 @@
-package org.infinispan.stats.topK;
+package org.infinispan.extendedstats.topK;
 
 import org.infinispan.configuration.cache.CacheMode;
-import org.infinispan.stats.BaseClusterTopKeyTest;
+import org.infinispan.extendedstats.BaseClusterTopKeyTest;
 import org.testng.annotations.Test;
 
 /**
  * @author Pedro Ruivo
  * @since 6.0
  */
-@Test(groups = "functional", testName = "stats.topK.ReplTopKeyTest")
+@Test(groups = "functional", testName = "extendedstats.topK.ReplTopKeyTest")
 public class ReplTopKeyTest extends BaseClusterTopKeyTest {
    public ReplTopKeyTest() {
       super(CacheMode.REPL_SYNC, 2);
