+++ extended-statistics/src/test/resources/log4j.xml
@@ -35,6 +35,10 @@
       <priority value="TRACE"/>
    </category>
 
+   <category name="org.infinispan.extendedstats">
+      <priority value="TRACE"/>
+   </category>
+
    <!-- ======================= -->
    <!-- Setup the Root category -->
    <!-- ======================= -->
